program1-trainer/requirements.txt
tensorflow==2.15.*
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.12.0
Pillow>=8.0.0
weasyprint>=54.0
PyYAML>=6.0